# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Mapa de Emergencia y Rescate — API",
    "version": "1.0.0",
    "description": "API pública/admin del mapa de emergencia. Documentación generada automáticamente desde los bloques @swagger de cada route."
  },
...
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "MissingPerson": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "age": {
            "type": "integer",
            "nullable": true
          },
          "description": {
            "type": "string"
          },
          "lastSeen": {
            "type": "string"
          },
          "contact": {
            "type": "string"
          },
          "photoUrl": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "found"
            ]
          },
          "resolutionNote": {
            "type": "string",
            "nullable": true
          },
          "resolutionPhotoUrl": {
            "type": "string",
            "nullable": true
          },
          "resolvedAt": {
            "type": "integer",
            "nullable": true
          },
          "createdAt": {
            "type": "integer",
            "description": "epoch-ms"
          }
        }
      },
...
      "Hospital": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "externalId": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "facilityType": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "municipality": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "level": {
            "type": "string",
            "nullable": true
          },
          "priorityZone": {
            "type": "string",
            "enum": [
              "P0",
              "P1",
              "P2",
              "P3"
            ]
          },
          "isPriority": {
            "type": "boolean"
          },
          "activePatients": {
            "type": "integer"
          },
          "totalPatients": {
            "type": "integer"
          },
          "createdAt": {
            "type": "integer"
          }
        }
      },
      "HospitalPatient": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "hospitalId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "age": {
            "type": "integer",
            "nullable": true
          },
          "condition": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "contact": {
            "type": "string"
          },
          "admittedAt": {
            "type": "integer"
          },
          "updatedAt": {
            "type": "integer"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/hospitals/{id}/patients": {
      "get": {
        "tags": [
          "hospitals"
        ],
        "summary": "Lista los pacientes de un hospital",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pacientes y datos del hospital",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "patients": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/HospitalPatient"
                      }
                    },
                    "hospital": {
                      "$ref": "#/components/schemas/Hospital"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Hospital no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "hospitals"
        ],
        "summary": "Registra un paciente en un hospital (rate-limited)",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "age": {
                    "type": "integer",
                    "nullable": true
                  },
                  "condition": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string"
                  },
                  "notes": {
                    "type": "string"
                  },
                  "contact": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Paciente creado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "patient": {
                      "$ref": "#/components/schemas/HospitalPatient"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos (nombre faltante o demasiado largo)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Hospital no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Demasiadas peticiones",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "No se pudo guardar el paciente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/missing/{id}/found": {
      "post": {
        "tags": [
          "missing"
        ],
        "summary": "Marca un reporte de persona desaparecida como localizada con nota y foto de prueba",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID del reporte de persona desaparecida"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "note",
                  "photo"
                ],
                "properties": {
                  "note": {
                    "type": "string",
                    "description": "Explicación de cómo se confirmó el contacto"
                  },
                  "photo": {
                    "type": "string",
                    "description": "Imagen (data URL JPG/PNG/WebP) como prueba del contacto"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Persona marcada como localizada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "person": {
                      "$ref": "#/components/schemas/MissingPerson"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Falta la nota o la foto, o son inválidas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "El reporte no existe o ya fue resuelto",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "La imagen excede el tamaño permitido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Demasiadas solicitudes (rate limit)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "No se pudo actualizar, intentar de nuevo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/missing/{id}/restore": {
      "post": {
        "tags": [
          "missing"
        ],
        "summary": "Restaura (admin) una persona desaparecida previamente marcada como localizada",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Restauración exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "No existe o no estaba marcada como localizada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}